refactor(web/engine): initial ModelManager split
