+++ web/source/osk/visualKeyboard.ts
@@ -907,7 +907,7 @@ namespace com.keyman.osk {
 
     getTouchProbabilities(touch: Touch): text.KeyDistribution {
       let keyman = com.keyman.singleton;
-      if(!keyman.modelManager.mayCorrect) {
+      if(!keyman.core.languageProcessor.mayCorrect) {
         return null;
       }
